[REFACTOR] 기능 주석 추가 및 용도에 맞게 refactoring #87
