--- src/main/java/com/example/namoldak/util/webSocket/ConnectHandler.java
@@ -1,4 +1,0 @@
-package com.example.namoldak.util.webSocket;
-
-public class ConnectHandler {
-}
